Adds show term function for localhost components
I added a few scripts that take care of creating, renaming and switching
to a 'clone' session.

Basically the start_named_clone_session checks if
a clone session for a component already exists and if not, it creates
one. By extracting the group number of the main session, the name of the
newly created clone session can be retrieved by assuming, this process
is the only process creating new clone sessions thus the session within
the master group with a number as name must be the newly created. After
retrieving its name, the session will be renamed after the
component_name-clone-session schema and its active window will be set to
the component window.

The show_term script opens up a new xterm which will execute the
show_session script. The former unsets its inherited $TMUX env var to be
able to start its own tmux session before running the attach command
(without this the main invoking terminal will be switched, because the
$TMUX env var is passed on to each subprocess.

These scripts are triggered by the "show term" buttons. Closing the term
on a detoggle is not implemented yet, but will come with one of the next
commits.

Also this procedure has to be adapted to handle the case, when a
component is running on another machine.

Resolves: #8