Fix unecessary new clone creation bug.
Checking the tagert window and the main session group where swapped in
the order, because on the group check a clone session might be started
to get a group assigned to the main session, which is pointless if the
target window does not exists and results in an unnamed clone session
which contradicts the assumption of the clone session creation process.

So now, if the target window does not exist, the script will exit
without creating a new session.