Fix master cmd execution in component windows bug
When sending commands to a session with libtmux, the currently active
window of the session will receive the command. Sending a command to the
main session without a specified window may have unwanted effects,
because if a component is started on the local machine, the active
window of the main session will switch to the new components one.

To prevent sending input to windows already executing a command, a new
funtion was introduced, to specially send input to the main window of
the main session.