Adds method for dependency resolution
The implemented approach creates a node for each component and, if
existing, adds a directed edge to specified component-dependencies. This
features missing dependency and circular dependency detection.
To check whether the configuration is valid, a pseudo node is
introduced, which depends on all components, thus if the recursive
function to resolve dependencies, a thrown circular dependency exception
indicates an error. Also after adding all components as nodes, a
dependency which is not among the nodes will be detected as missing.