Introduces viusal validation of configs
Dependency visualisation:
Using the nodes and edges of the dependency resolution, a directed graph from
the graphviz library is created, which will display components with
dependencies as connected nodes. A missing dependency is displayed as red node
and the arrow pointing to it is also red and labeled with 'missing'. A
circular dependency error can be displayed too, but due to the error the
rest of the graph can not be iterated, so unless the exception occurs on
the last two nodes, the displayed graph will be incomplete.