Adds component check feature for localhost
On check the window is searched in the session, if it is not found and
no custom check was specified, false is returned. If the check succeeds
the tool knows, that the component is running, but was not started with
hyperion.
If the window does exsist, the window child processes are queried. If
the amount is not correct and no check is available, false will be
returned. If the pid number checks out, a custom check will be run if
available, else true is returned.

For now no extra thread is used. If the check takes longer, the gui will
freeze.

refs: #12