Adds remote show term feature.
This commit introduces a TMUX section in the Control Center class.
The clone session subprocess call from the gui was moved to a function
in the Control Center class and the command in run in the master window
of the master session.
Also a kill_remote_session_by_name and a start_remote_clone_session
method have been implemented (also both executing the commands  in the
master session).

Resolves: #9