Introduce cli single command run feature
This commit splits the previously called run mode into two separate call
options: the gui mode and the cli mode.
The gui mode works like the run mode used to before, initializing the
main session, copying components to remote hosts if needed and starts
the GUI.

The cli mode provides the possibility to start, stop or check a single,
a set or all components over the command line interface, as the name
suggests. Additionally an option to list all components of the current
setup is featured.