Introduce using ssh master connection
On system startup the users ssh config is copied to a temporary file and
options are added to allow master connections. After that, wenn parsing
the config, for each master, that is reachable a master connection is
established (meaning no new tcp connections on ssh commands to the same
host have to be opened).

Each master connection is run in a component-like window of the master
session. Also a cleanup function was added, to end those connections on
application shutdown. Finally also the master session is killed on exit.