Introduce first process monitoring prototype
The init starts up a thread that handles a monitoring job queue which is
filled by the main thread. Two kinds of jobs can be added to the queue:
process monitoring jobs or host monitoring jobs. The worker thread runs
their implicitly defined checks (checking if a pid is still running) and
passes on a generated event (disconnect or crash) to all subscribers or,
on a successful check puts the jobs back in the queue for the next
iteration.

The monitor thread is killed on application exit in the cleanup section.

On initialisation of a master ssh connection, the pid of the ssh command
(extracted as child from the executing tmux window process) is added to
the queue as monitoring job. This opens up the possibility of handling
connection losses.

Right now no remote pid check is supported.

Refs: #16