Fix check_local_component function
The function now returns a tuple of the component checkstate and a pid.
If the component is running, the pid of the first child process that is
not a tee logging process. If the component is not running or the window
running or the components tmux widnow does not exist (even if the check
succeeds), the returned pid will be '0'.