Implement new check behaivor for slave and master
On a slave component check the pid is printed to the commandline and the
CheckState enum value is returned as system return value. A master
session performing a remote check opens an ssh connection to the remote
host and runs a slave in check mode. When the slave process has finished
the pid is parsed from the command line and the return status is
interpreted as CheckState, which then is passed on.

For the parsed pid (if it is not the fake pid 0) a new monitoring job is
created, which then is fed into the monitoring queue to handle a
component crash.