Add basic monitoring library
The library introduces thread based monitoring with a main thread
checking a job queue which contains newly defined host or component
monitoring jobs, that have their checks and information implicitly
defined. This means the thread runs the predefined checks and on success
puts them back into the queue to rerun the check in the next iteration.
On fail a crash event (again implicitly created by the job itself) is
put into each subscriber queue.
Registering a subscriber queue is of course provided by the thread
object.

The main idea is, that this non-busy thread handles component an host
monitoring in terms of detecting a crash and notifying all subscribers,
that is all interactive interfaces like the gui or the interactive cli.

Remote checks are not yet supported but all needed information is now
available to the monitoring thread.

Refs: #16