Split process monitoring in more classes
The process monitoring class has been split into a local and a remote
monitoring class. The same happened to the crash event class.

Functionality for a remote host check has been implemented: a ssh
connection is opened and a ps specifically requesting the remote pid is
executed. This check is only executed though, if the remote host is
reachable.

The main monitoring thread now sorts the jobs in its queue to first
process host monitoring and then component monitoring. This prevents
firing a lot of crash events, if the connection of a host is lost. The
processes of a host that the connection was lost to are simply set back
into the queue to process as soon as the connection is restored.

Also docstring was added for threads.py

Refs: #16, #17