Patch ping command execution
When ping is executed over os.system input from the terminal may be
forwarded to the ping command (so if C-c is typed the SIGINT may be sent
to the ping instead of the main application). To prevent this, this
commit nests the ping command in an exec command and sets its stdin to
/dev/null. This way the ping command won't recieve any input from the
invoking tty.