Patch host connection status detection
os.system most likely opens /bin/sh which does not support redirecting
in the "exec > >(...)" command. Because of this the command was again
changed but now does not have a stdin redirection because getting the
return status was getting too complicated. The error this change was
made for (ctrl-c may be forwarded to the ping command) is solved for GUI
mode because of the SIGINT -> SIG_DFL connection. I hope this will for
the interactive CLI too.