Add function to wait for window availability
This function checks whether the supplied window has child processes
left, that are still running and waits until all of them are finished.
This makes a basic kind of scheduling of window commands available to
prevent window commands interrupting previously invoked commands.

The additional wait for main window function fetches the main window of
the master session and invokes the above function.

Fixes: #13