Adds reusing already running ssh connections
Introduces check if window is busy (moved from wait for window not busy
function) this is used to check if an ssh connection is still active in
an old window. The downside to this is, that if a hostname was changed
after the first connection was established, the system will not see the
need of reconnecting because host resolution of the running connection
is not a part of the check.