Configure logger with new config file
The logger is now configured with a file contained in the /data
directory. New file handlers have been added: the full DEBUG level log
is stored in /tmp/Hyperion/log/debug.log, the INFO level log is
stored in /tmp/Hyperion/log/info.log and a screen handler prints DEBUG
level messages to stdout.

Also adaptive handler config is introduced:
When cli mode is started the screen handler log level is set to INFO and
if interactive mode was chosen, the screen handler is discarded
(prints would appear in the ui).