Fix using old window for new component start
When an old window was used for restarting a component, all its child
processes were killed, including the tee logging processes. This causes
a SIGPIPE signal on the executing process as soon as it gets new input,
because it wants to write to a pipe with no readers resulting in the
termination of the window process.
The solution to this problem is to start the tee processes once per
window and in case they already exist only rotate logs before issuing
the components start command.

The old logfile is now renamed to <Component Name>-<Date Time>.log to
make it easier finding recent logs.