I-CLI: Add option to hide application log
This option is helpful because the log will always show 10 rows and is
always visible, beacuse it is located in the main frame footer. If the
window is small, the footer may be filling the whole window. In this
case the log can be hidden to make room for the frame body.