I-CLI: Add shutdown popup with user choice
On 'Q' press the shutdown menu popup is displayed, which can be
dismissed by choosing the cancel button or pressing 'esc'. The popup
offers the choice between a full scale shutdown or just closing the
interface.

When urwid is closed by raising urwid.ExitMainLoop(), the main_loop()
function in interactiveCLI exits. A line after was added, returning the
full_shutdown bool of the UI to the runner thread which is passed on to
the cleanup function of the ControlCenter.