Add serializer for requests & responses via socket
The serializer receives an action type (str) and a payload (list of
objects) which is encoded in a dictionary and pickled. The length of the
full message is added as first 4 bytes of the message.
The deserializer unpickles and decodes a serialized message and returns
action type and argument list as tuple. The deserializer expects that
the leading 4 bytes with lenght information have been removed before.

Refs: #23