Adds basic server and remoteClient classes
Server wraps around the hyperion core and remoteClient wraps around a
ui. Both act as an interface for communication and use the
actionSerializer and sockets for communication.
The philosophy of the remoteClient is to keep calls of a ui the same as
if it would run in standalone mode, meaning each callable function or
    accessible member of the hyperion core needs to have a counterpart
    in the remoteClient class (either resulting in a message to the
    server or returning information gained by previous communication
    with the server).

Currently only config and host_list are received by the client, but the
basic communication works and is easily extendable.

Refs: #23