Introduce fully async response model in manager
The manager now handles starting a component with its dependencies. The
simple reason behind this is a strongly reduced communication between
the manager and a ui (or better yet reduction to async communication
only). The manager now handles an own subscriber queue to broadcast
events to. Via the new _broadcast_event function it is ensured, that all
subscribed clients will be informed about component starts, checks and
stops. All sendable events are stored in lib.util.events.

Refs: #23