Fix server host_list request processing
Trying to serialize a host list dictionary that includes remote hosts
fails due to the fact that it contains psutil.Process objects that can't
be pickled by default. Given the fact, that a ui does not need to know
the process id of the ssh master connection, the dict that gets sent
know only contains a boolean.