Removed pinging remotes from mon thread
Added ServerAliveInterval to custom ssh config to detect disconnect
(every ten seconds). On disconnect the ssh process will terminate which
now is monitored by the mon thread to discover disconnects. On socket
write attempts to the slave it will result in a unreachable state
return. Again, as mentioned in #29 a new display state for hosts could
be added to indicate the slave connection was lost, but ssh is still
running (could indicate a missing hyperion installation or signal
something is wrong with the connection but ssh did not notice yet).