Introduce force mode as core feature
Starting components is really strict now: in normal mode, all
dependencies of a component are resolved and each is tried to start. If
a single component fails, all following components will only be checked
and if they are not running, be marked with `DEPENDENCY_FAILED`.
In force mode every component start is attempted regardless of prevoius
failures.

Resolves #33