Fixes several minor issues
* Adds utility to check if a local component is already being monitored.
* Moved component monitor creation to common AbstractController parent
class.
* Replaces _get_window_pid function with libtmux's  window.pane_pid
property. This saves time.
* Replaces checking busy windows with checking which program they are
running. If it is a shell, window is idling. This is much faster,
because it avoids costly calls to psutil process children search.
* Checking a component now first checks if a component is already being
monitored. If it is, true is returned, because this means that a
previous check was successful and the process is still running. With
this approach only a single constly psutil call to get the process to be
monitored is needed, which saves a lot of time in repeated checks.
* Adds extra single threaded versions if nthreads is set to 1. This
avoids the overhead of threadpool creation and management.
* Fixed checks of the final component batch being omitted when stopping
all components.